- VS leave message for smooth leaving - Introduced byte_t to make ReadBuf ctor more typesafe - Fixed readbuf offset bug in EVSProto::resend() (and inserted FIXME into recover()) - Get representative method in view - Nil UUID - Safer ReadBuf implementation for debugging - Some EVS proto fixes, setting operational flag properly
Bugs still: There still seems to be a bug in trans msg delivery which causes message gap exception on VS level




git-svn-id: https://devel.codership.com/svn/galera/trunk@773 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ gcomm/src/gmcast.hpp
@@ -280,7 +280,7 @@ public:
             return 0;
         if ((off = source_uuid.read(buf, buflen, off)) == 0)
             return 0;
-
+        
         if (flags & F_NODE_ADDRESS)
         {
             char* addr = 0;
@@ -313,9 +313,9 @@ public:
                 node_list->push_back(node);
             }
         }
-        LOG_DEBUG(UInt8(type).to_string() + "," 
-                  + UInt8(flags).to_string() + "," 
-                  + UInt8(ttl).to_string() + ","
+        LOG_DEBUG(Int(type).to_string() + "," 
+                  + Int(flags).to_string() + "," 
+                  + Int(ttl).to_string() + ","
                   + (F_GROUP_NAME ? group_name : ""));
         return off;
     }
